fix(android): serialize share cleanup recovery
