+++ androidApp/src/main/kotlin/dev/obiente/nextcloudnative/AndroidIncomingShareUploads.kt
@@ -1,44 +1,27 @@
 package dev.obiente.nextcloudnative
 
-import android.Manifest
-import android.app.ForegroundServiceStartNotAllowedException
-import android.app.Notification
-import android.app.PendingIntent
 import android.content.ContentResolver
 import android.content.Context
 import android.content.Intent
-import android.content.pm.PackageManager
-import android.content.pm.ServiceInfo
 import android.database.Cursor
 import android.net.Uri
-import android.os.Build
 import android.provider.OpenableColumns
 import android.util.AtomicFile
-import androidx.annotation.RequiresApi
-import androidx.core.app.NotificationCompat
-import androidx.core.app.NotificationManagerCompat
-import androidx.core.content.ContextCompat
-import androidx.work.CoroutineWorker
-import androidx.work.ForegroundInfo
-import androidx.work.WorkerParameters
 import dev.obiente.nextcloudnative.app.MAX_INCOMING_SHARE_FILES
-import dev.obiente.nextcloudnative.app.NextcloudSession
 import dev.obiente.nextcloudnative.app.canonicalIncomingShareDestinationPath
 import dev.obiente.nextcloudnative.app.incomingShareUploadNameCandidates
 import dev.obiente.nextcloudnative.app.safeIncomingShareFileName
-import dev.obiente.nextcloudnative.app.useAndroidNextcloudCertificateTrust
 import java.io.File
 import java.io.FileOutputStream
 import java.util.UUID
-import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.currentCoroutineContext
 import kotlinx.coroutines.ensureActive
-import kotlinx.coroutines.job
 import kotlinx.coroutines.withContext
-import okhttp3.OkHttpClient
 import org.json.JSONArray
 import org.json.JSONObject
+
+internal const val ABANDONED_INCOMING_SHARE_STAGING_RETENTION_MILLIS = 24L * 60L * 60L * 1_000L
 
 internal enum class AndroidIncomingShareState {
     Staged,
@@ -117,6 +100,7 @@ internal class AndroidIncomingShareStore(private val context: Context) {
         val requestId = UUID.randomUUID().toString()
         val requestDirectory = directory(requestId)
         check(requestDirectory.mkdirs()) { "The private upload staging folder could not be created." }
+        scheduleIncomingShareAbandonedStagingCleanup(context, requestId)
         val stagingMarker = createIncomingShareStagingMarker(requestDirectory, STAGING_MARKER_NAME)
         try {
             var totalBytes = 0L
@@ -193,7 +177,7 @@ internal class AndroidIncomingShareStore(private val context: Context) {
 
     fun listRecoverable(accountId: String): List<AndroidIncomingShareRequest> = synchronized(LOCK) {
         removeExpiredAbandonedIncomingShareStaging(
-            root, STAGING_MARKER_NAME, ABANDONED_STAGING_RETENTION_MILLIS,
+            root, STAGING_MARKER_NAME, ABANDONED_INCOMING_SHARE_STAGING_RETENTION_MILLIS,
         )
         root.listFiles().orEmpty()
             .asSequence()
@@ -223,6 +207,22 @@ internal class AndroidIncomingShareStore(private val context: Context) {
             atomic.failWrite(stream)
             throw failure
         }
+    }
+
+    fun queue(
+        id: String,
+        accountId: String,
+        userId: String,
+        destinationPath: String,
+    ): AndroidIncomingShareRequest = synchronized(LOCK) {
+        val queued = prepareIncomingShareRequestForQueue(
+            current = requireAvailable(id),
+            accountId = accountId,
+            userId = userId,
+            destinationPath = destinationPath,
+        )
+        save(queued)
+        queued
     }
 
     fun transition(
@@ -327,6 +327,24 @@ internal class AndroidIncomingShareStore(private val context: Context) {
         updated
     }
 
+    fun claimChunkSessionForCleanup(
+        id: String,
+        uploadId: String,
+        includeRetryableFailure: Boolean,
+    ): AndroidIncomingShareChunkSession? = synchronized(LOCK) {
+        val current = load(id) ?: return@synchronized null
+        val session = current.chunkSession?.takeIf { it.uploadId == uploadId }
+            ?: return@synchronized null
+        if (current.state !in TERMINAL_INCOMING_SHARE_STATES) return@synchronized null
+        val eligible = session.cleanupPending ||
+            current.state != AndroidIncomingShareState.Failed ||
+            includeRetryableFailure
+        if (!eligible) return@synchronized null
+        val claimed = session.copy(commitInFlight = false, cleanupPending = true)
+        if (claimed != session) save(current.copy(chunkSession = claimed))
+        claimed
+    }
+
     fun clearChunkSession(id: String): AndroidIncomingShareRequest = synchronized(LOCK) {
         val current = requireAvailable(id)
         require(current.state == AndroidIncomingShareState.Uploading)
@@ -348,6 +366,17 @@ internal class AndroidIncomingShareStore(private val context: Context) {
         target.isDirectory && target.deleteRecursively()
     }
 
+    fun removeExpiredAbandonedStaging(id: String, nowMillis: Long = System.currentTimeMillis()): Boolean =
+        synchronized(LOCK) {
+            val target = directory(id)
+            !target.exists() || removeExpiredAbandonedIncomingShareStagingDirectory(
+                directory = target,
+                markerName = STAGING_MARKER_NAME,
+                retentionMillis = ABANDONED_INCOMING_SHARE_STAGING_RETENTION_MILLIS,
+                nowMillis = nowMillis,
+            )
+        }
+
     fun removeStagedFiles(request: AndroidIncomingShareRequest) = synchronized(LOCK) {
         request.files.forEach { file ->
             val staged = File(directory(request.id), file.stagedName)
@@ -367,252 +396,8 @@ internal class AndroidIncomingShareStore(private val context: Context) {
         const val MAX_SHARE_FILE_BYTES = 8L * 1024L * 1024L * 1024L
         const val MAX_SHARE_TOTAL_BYTES = 16L * 1024L * 1024L * 1024L
         const val MIN_STAGING_FREE_BYTES = 64L * 1024L * 1024L
-        const val ABANDONED_STAGING_RETENTION_MILLIS = 24L * 60L * 60L * 1_000L
         const val STAGING_MARKER_NAME = ".staging"
         const val MAX_INCOMING_SHARE_RECOVERY_SCAN = 1_000
-    }
-}
-
-internal class AndroidIncomingShareUploadWorker(
-    appContext: Context,
-    params: WorkerParameters,
-) : CoroutineWorker(appContext, params) {
-    override suspend fun doWork(): Result = withContext(Dispatchers.IO) {
-        val requestId = inputData.getString(KEY_REQUEST_ID) ?: return@withContext Result.failure()
-        val store = AndroidIncomingShareStore(applicationContext)
-        var request = when (val loaded = store.loadResult(requestId)) {
-            is AndroidIncomingShareLoadResult.Available -> loaded.request
-            is AndroidIncomingShareLoadResult.Corrupt -> {
-                publishCorruptIncomingShareNotification(applicationContext, loaded.requestId)
-                scheduleIncomingShareCleanup(applicationContext, loaded.requestId)
-                return@withContext Result.failure()
-            }
-            AndroidIncomingShareLoadResult.Missing -> return@withContext Result.success()
-        }
-        if (request.state == AndroidIncomingShareState.Uploading) {
-            val resumable = request.chunkSession?.takeIf { !it.commitInFlight }
-            if (resumable != null) {
-                request = store.transition(
-                    id = requestId,
-                    expected = setOf(AndroidIncomingShareState.Uploading),
-                    target = AndroidIncomingShareState.Queued,
-                    message = "Resuming the large file from its last saved chunk.",
-                ) ?: return@withContext Result.success()
-            } else {
-            val recovered = store.transition(
-                id = requestId,
-                expected = setOf(AndroidIncomingShareState.Uploading),
-                target = AndroidIncomingShareState.OutcomeUnknown,
-                message = "Android restarted during an upload. Check Files before trying again.",
-            )
-            recovered?.let {
-                publishTerminalNotification(it)
-                scheduleIncomingShareCleanup(applicationContext, it.id)
-            }
-            return@withContext Result.success()
-            }
-        }
-        if (request.state != AndroidIncomingShareState.Queued) return@withContext Result.success()
-        val session = AndroidNextcloudServices(applicationContext).loadSession()
-        if (session == null || NextcloudDocumentIds.accountKey(session) != request.accountId) {
-            val failed = store.transition(
-                id = requestId,
-                expected = setOf(AndroidIncomingShareState.Queued),
-                target = AndroidIncomingShareState.Failed,
-                message = "The upload account is not active.",
-            )
-            failed?.let {
-                publishTerminalNotification(it)
-                scheduleIncomingShareCleanup(applicationContext, it.id)
-            }
-            return@withContext Result.failure()
-        }
-        AndroidNotificationCoordinator(applicationContext).ensureChannels()
-        runCatching { setForeground(foregroundInfo(request)) }
-            .onFailure { failure ->
-                if (failure !is IllegalStateException || !isForegroundStartUnavailable(failure)) throw failure
-            }
-        request = store.transition(
-            id = requestId,
-            expected = setOf(AndroidIncomingShareState.Queued),
-            target = AndroidIncomingShareState.Uploading,
-        ) ?: return@withContext Result.success()
-        val remote = AndroidFileSyncRemoteTree(
-            session = session,
-            userId = requireNotNull(request.userId),
-            remoteRootPath = requireNotNull(request.destinationPath),
-            webDav = NextcloudDocumentWebDav(
-                client = OkHttpClient.Builder()
-                    .followRedirects(false)
-                    .followSslRedirects(false)
-                    .useAndroidNextcloudCertificateTrust(applicationContext)
-                    .build(),
-                cloudMutationsAllowed = applicationContext.cloudMutationGate(),
-            ),
-        )
-        val requestCancellation = CoroutineDocumentRequestCancellation(currentCoroutineContext().job)
-        var mutationInFlight = false
-        try {
-            val occupiedNames = remote.rootChildNames().names.toMutableSet().apply {
-                addAll(request.uploadedNames)
-            }
-            val transfer = AndroidIncomingShareFileTransfer(store, remote, requestCancellation)
-            for (index in request.completedFiles until request.files.size) {
-                ensureNotCanceled(requestId, store)
-                request = transfer.upload(requestId, request, index, occupiedNames) { inFlight ->
-                    mutationInFlight = inFlight
-                }
-                setForeground(foregroundInfo(request))
-            }
-            scheduleIncomingShareCleanup(applicationContext, request.id)
-            request = store.transition(
-                id = requestId,
-                expected = setOf(AndroidIncomingShareState.Uploading),
-                target = AndroidIncomingShareState.Completed,
-            ) ?: throw CancellationException("Incoming share upload canceled")
-            runCatching { store.removeStagedFiles(request) }
-            publishTerminalNotification(request)
-            Result.success()
-        } catch (cancelled: CancellationException) {
-            if (store.load(requestId)?.state != AndroidIncomingShareState.Canceled) {
-                val transitioned = store.transition(
-                    id = requestId,
-                    expected = setOf(AndroidIncomingShareState.Uploading),
-                    target = if (mutationInFlight) {
-                        AndroidIncomingShareState.OutcomeUnknown
-                    } else {
-                        AndroidIncomingShareState.Queued
-                    },
-                    message = if (mutationInFlight) {
-                        "Android stopped during an upload. Check Files before trying again."
-                    } else {
-                        "Upload paused. It will continue when Android allows background work."
-                    },
-                )
-                transitioned?.takeIf { it.state == AndroidIncomingShareState.OutcomeUnknown }?.let {
-                    publishTerminalNotification(it)
-                    scheduleIncomingShareCleanup(applicationContext, it.id)
-                }
-            }
-            throw cancelled
-        } catch (failure: Throwable) {
-            val definitelyRejected = mutationInFlight && !incomingShareMutationOutcomeUnknown(failure, true)
-            if (definitelyRejected) {
-                store.load(requestId)?.chunkSession?.takeIf { it.commitInFlight }?.let {
-                    store.clearChunkCommitInFlight(requestId)
-                }
-                mutationInFlight = false
-            }
-            val resumableChunk = store.load(requestId)?.chunkSession?.takeIf { !it.commitInFlight }
-            val retryable = !mutationInFlight &&
-                failure.isRetryableIncomingShareTransferFailure() &&
-                runAttemptCount + 1 < MAX_INCOMING_SHARE_TRANSFER_ATTEMPTS
-            if (retryable || resumableChunk != null && runAttemptCount + 1 < MAX_INCOMING_SHARE_TRANSFER_ATTEMPTS) {
-                store.transition(
-                    id = requestId,
-                    expected = setOf(AndroidIncomingShareState.Uploading),
-                    target = AndroidIncomingShareState.Queued,
-                    message = "Upload paused and will retry with backoff.",
-                )
-                return@withContext Result.retry()
-            }
-            // A transport failure after a conditional PUT starts cannot prove whether the server
-            // committed it. Do not replay automatically and risk a duplicate.
-            val outcomeUnknown = incomingShareMutationOutcomeUnknown(failure, mutationInFlight)
-            val target = if (outcomeUnknown) {
-                AndroidIncomingShareState.OutcomeUnknown
-            } else {
-                AndroidIncomingShareState.Failed
-            }
-            val transitioned = store.transition(
-                id = requestId,
-                expected = setOf(AndroidIncomingShareState.Uploading),
-                target = target,
-                message = failure.message?.take(240) ?: if (outcomeUnknown) {
-                    "The upload result is unknown."
-                } else {
-                    "The upload could not continue."
-                },
-            )
-            transitioned?.let {
-                publishTerminalNotification(it)
-                scheduleIncomingShareCleanup(applicationContext, it.id)
-            }
-            Result.failure()
-        } finally {
-            requestCancellation.close()
-        }
-    }
-
-    private fun ensureNotCanceled(requestId: String, store: AndroidIncomingShareStore) {
-        if (store.load(requestId)?.state == AndroidIncomingShareState.Canceled) {
-            throw CancellationException("Incoming share upload canceled")
-        }
-    }
-
-    private fun foregroundInfo(request: AndroidIncomingShareRequest): ForegroundInfo {
-        val notification: Notification = NotificationCompat.Builder(applicationContext, CHANNEL_TRANSFERS)
-            .setSmallIcon(R.drawable.ic_notification)
-            .setColor(0xFF8F5EAD.toInt())
-            .setContentTitle("Uploading shared files")
-            .setContentText("${request.completedFiles} of ${request.files.size} uploaded")
-            .setCategory(NotificationCompat.CATEGORY_PROGRESS)
-            .setPriority(NotificationCompat.PRIORITY_LOW)
-            .setOngoing(true)
-            .setOnlyAlertOnce(true)
-            .setProgress(request.files.size, request.completedFiles, false)
-            .setContentIntent(incomingShareRecoveryPendingIntent(applicationContext, request.id))
-            .build()
-        val id = request.id.hashCode().let { if (it == Int.MIN_VALUE) 1 else kotlin.math.abs(it) }.coerceAtLeast(1)
-        return if (Build.VERSION.SDK_INT >= 29) {
-            ForegroundInfo(id, notification, ServiceInfo.FOREGROUND_SERVICE_TYPE_DATA_SYNC)
-        } else {
-            ForegroundInfo(id, notification)
-        }
-    }
-
-    private fun isForegroundStartUnavailable(error: IllegalStateException): Boolean =
-        Build.VERSION.SDK_INT >= 31 && isForegroundStartUnavailableApi31(error)
-
-    @RequiresApi(31)
-    private fun isForegroundStartUnavailableApi31(error: IllegalStateException): Boolean =
-        error is ForegroundServiceStartNotAllowedException
-
-    private fun publishTerminalNotification(request: AndroidIncomingShareRequest) {
-        if (
-            Build.VERSION.SDK_INT >= 33 &&
-            ContextCompat.checkSelfPermission(applicationContext, Manifest.permission.POST_NOTIFICATIONS) !=
-            PackageManager.PERMISSION_GRANTED
-        ) return
-        AndroidNotificationCoordinator(applicationContext).ensureChannels()
-        val completed = request.state == AndroidIncomingShareState.Completed
-        val notification = NotificationCompat.Builder(applicationContext, CHANNEL_TRANSFERS)
-            .setSmallIcon(R.drawable.ic_notification)
-            .setColor(0xFF8F5EAD.toInt())
-            .setContentTitle(if (completed) "Shared files uploaded" else "Shared upload needs attention")
-            .setContentText(
-                if (completed) {
-                    "${request.completedFiles} files uploaded to Nextcloud"
-                } else {
-                    "${request.completedFiles} of ${request.files.size} uploaded. Tap to review."
-                },
-            )
-            .setCategory(if (completed) NotificationCompat.CATEGORY_STATUS else NotificationCompat.CATEGORY_ERROR)
-            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
-            .setAutoCancel(true)
-            .setOnlyAlertOnce(true)
-            .setContentIntent(incomingShareRecoveryPendingIntent(applicationContext, request.id))
-            .build()
-        try {
-            NotificationManagerCompat.from(applicationContext)
-                .notify(incomingShareNotificationId(request.id), notification)
-        } catch (_: SecurityException) {
-            // The permission can still be revoked between the explicit check and delivery.
-        }
-    }
-
-    internal companion object {
-        const val KEY_REQUEST_ID = "request_id"
     }
 }
 
@@ -637,6 +422,9 @@ internal fun prepareIncomingShareRequestForQueue(
     require(current.state == AndroidIncomingShareState.Staged || current.state == AndroidIncomingShareState.Failed) {
         "This upload is no longer waiting to be queued."
     }
+    require(current.chunkSession?.cleanupPending != true) {
+        "Nextcloud is still cleaning up the previous upload attempt. Try again shortly."
+    }
     require(accountId.isNotBlank() && userId.isNotBlank())
     val destination = canonicalIncomingShareDestinationPath(destinationPath)
     require((current.completedFiles == 0 && current.chunkSession == null) || current.destinationPath == destination) {
@@ -650,6 +438,11 @@ internal fun prepareIncomingShareRequestForQueue(
         message = null,
     )
 }
+
+internal fun AndroidIncomingShareRequest.isFullyJournaledIncomingShareUpload(): Boolean =
+    state == AndroidIncomingShareState.Uploading &&
+        completedFiles == files.size &&
+        chunkSession == null
 
 @Suppress("DEPRECATION")
 internal fun incomingShareUris(intent: Intent): List<Uri> {
